==> code/06_figures.qmd <==
---
title: |
  Code Notebook
subtitle: |
  How does reported partisan attachement influence affective polarization?: A comparative study of 25 European democracies
author:
  - name: Tristan Muno
    email: tristan.muno@uni-mannheim.de
    affiliations:
      - name: University of Mannheim
        department: School of Social Sciences
date: last-modified
date-format: MMMM D, YYYY
format:
  html:
    toc: true
    code-fold: true
    code-tools: true
    embed-resources: false
execute:
  echo: true
  warning: true
  eval: true
  message: true
---

# Setup

```{r}
#| label: setup

# To track render duration
start_time <- Sys.time()

# set width of console output
# options(width = 80)

# Install and load required packages
p_required <- c(
  "tidyverse",
  "here",
  "ggpubr",
  "gridExtra",
  "sessioninfo"
)
packages <- rownames(installed.packages())
p_to_install <- p_required[!(p_required %in% packages)]
if (length(p_to_install) > 0) {
  install.packages(p_to_install)
}
sapply(p_required, require, character.only = TRUE)
rm(p_required, p_to_install, packages)
```

# Load plot files

```{r}
#| label: load-models

load(here("data", "03_final", "plot_base_d.RData"))

load(here("data", "03_final", "plot_base_t.RData"))

load(here("data", "03_final", "plot_cov_d.RData"))

load(here("data", "03_final", "plot_cov_t.RData"))


load(here("data", "03_final", "plot_country_base_d.RData"))

load(here("data", "03_final", "plot_country_base_t.RData"))

load(here("data", "03_final", "plot_country_cov_d.RData"))

load(here("data", "03_final", "plot_country_cov_t.RData"))
```

# Figures

## Base dictator game model

```{r}
#| label: fig-pooled-base-d
#| fig-cap: "Combined figure"
#| fig-dpi: 500
#| fig-width: 6.3
#| fig-height: 9

ggarrange(
  plot_base_d[["AP_Plot"]],
  plot_base_d[["IF_Plot"]],
  plot_base_d[["OD_Plot"]],
  nrow = 3
)

```

## Base trust game model

```{r}
#| label: fig-pooled-base-t
#| fig-cap: "Combined figure"
#| fig-dpi: 500
#| fig-width: 6.3
#| fig-height: 9

ggarrange(
  plot_base_t[["AP_Plot"]],
  plot_base_t[["IF_Plot"]],
  plot_base_t[["OD_Plot"]],
  nrow = 3
)

```

## Covariate dictator model

```{r}
#| label: fig-pooled-cov-d
#| fig-cap: "Combined figure"
#| fig-dpi: 500
#| fig-width: 6.3
#| fig-height: 9

ggarrange(
  plot_cov_d[["AP_Plot"]],
  plot_cov_d[["IF_Plot"]],
  plot_cov_d[["OD_Plot"]],
  nrow = 3
)

```

## Covariate trust model

```{r}
#| label: fig-pooled-cov-t
#| fig-cap: "Combined figure"
#| fig-dpi: 500
#| fig-width: 6.3
#| fig-height: 9

ggarrange(
  plot_cov_t[["AP_Plot"]],
  plot_cov_t[["IF_Plot"]],
  plot_cov_t[["OD_Plot"]],
  nrow = 3
)

```

### Country figures

## Base dictator game model

```{r}
#| label: fig-country-base-d
#| fig-cap: "Combined figure"
#| fig-dpi: 500
#| fig-width: 6.3
#| fig-height: 9

plot_country_base_d

```

## Base trust game model

```{r}
#| label: fig-country-base-t
#| fig-cap: "Combined figure"
#| fig-dpi: 500
#| fig-width: 6.3
#| fig-height: 9

plot_country_base_t

```

## Covariate dictator model

```{r}
#| label: fig-country-cov-d
#| fig-cap: "Combined figure"
#| fig-dpi: 500
#| fig-width: 6.3
#| fig-height: 9

plot_country_cov_d

```

## Covariate trust model

```{r}
#| label: fig-country-cov-t
#| fig-cap: "Combined figure"
#| fig-dpi: 500
#| fig-width: 6.3
#| fig-height: 9

plot_country_cov_t

```

# Session Info

```{r}
#| label: session-info

session_info()
```


# Render Time

```{r}
#| label: render-time

end_time <- Sys.time()

rendering_time <- end_time - start_time

message(paste(
  "Document rendered in:",
  round(as.numeric(rendering_time, units = "secs"), 2),
  "seconds.
"
))
```
